% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dose_response.R
\name{analyze_data_dose_response}
\alias{analyze_data_dose_response}
\title{Fit dose-response curves to individual peptide replicates}
\usage{
analyze_data_dose_response(
  data_processed,
  c0.as = log10(1e-12),
  R2.threshold = 0.9,
  min_rep_hits = 3,
  keep.cols,
  multicore = T,
  silent = F
)
}
\arguments{
\item{data_processed}{output list or tibble from process_data_dose_response()
or prefilter_data_dose_response()}

\item{c0.as}{replacement for log10(0)}

\item{R2.threshold}{threshold for individual curves to be accepted}

\item{min_rep_hits}{minimum number of valid fitted curves per peptide}

\item{keep.cols}{Which columns should be copied from data_raw to 
data_peptides}

\item{multicore}{should curve fitting be parallelized (T for automatic 
detection of number of cores to use; number to specific manually; F to 
disable)}

\item{silent}{Turn off feedback from the function?}
}
\value{
list of EC50 results at peptide ($data_peptides) and replicates 
($data_replicates) level
}
\description{
Fit dose-response curves to individual peptide replicates
}
\examples{
\dontrun{
data_EC50 <- analyze_data_dose_response(data_processed,
  c0.as = log10(1e-12),
  R2.threshold = 0.9,
  min_rep_hits = 3,
  keep.cols = c("Protein.Group",
    "Protein.Ids",
    "Protein.Names",
    "Genes",
    "First.Protein.Description"),
  multicore = T,
  silent = F)
}
}
