% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dose_response.R
\name{analyze_data_dose_response_concentrations}
\alias{analyze_data_dose_response_concentrations}
\title{Analyze individual concentrations of dose-response experiment as groups}
\usage{
analyze_data_dose_response_concentrations(
  data_processed,
  filter_rows,
  min_fraction_per_group = 1,
  norm.method = "none",
  p.adjust.method = "BH",
  p.threshold = 0.01,
  fc.threshold = log2(1.2),
  eBayes.trend = TRUE,
  keep.cols = c(),
  rename.output = c(estimate = "log2.fc"),
  full.output = F,
  silent = T,
  translate.concentrations = T
)
}
\arguments{
\item{data_processed}{output list or tibble from process_data_dose_response()
or prefilter_data_dose_response()}

\item{filter_rows}{tidy expression to filter the data e.g. filter_rows = 
Lib.Q.Value < 0.01}

\item{min_fraction_per_group}{minimum fraction of values per replicate to use 
(e.g. 0.75 = min. 3 of 4 replicates with valid value)}

\item{norm.method}{which method to use for normalization}

\item{p.adjust.method}{method for p.adjust() (see p.adjust.methods)}

\item{p.threshold}{p-value threshold after adjustment}

\item{fc.threshold}{log2. fold-change threshold}

\item{eBayes.trend}{abundance-dependent variance estimation in the 
limma::eBayes function (default = TRUE)}

\item{keep.cols}{Which columns should be copied from data_raw to 
data_peptides}

\item{rename.output}{named vector of columns to rename in the output (e.g. 
#' c(name = "new_name"))}

\item{full.output}{Include full limma data in output?}

\item{silent}{Turn off feedback from the function?}

\item{translate.concentrations}{replace numeric concentrations with strings 
with unit}
}
\value{
list of concentration-grouped list results at peptide ($data_peptides) and replicates 
($data_replicates) level
}
\description{
Analyze individual concentrations of dose-response experiment as groups
}
\examples{
\dontrun{
data_limma <- analyze_data_dose_response_concentrations(data_processed,
  min_fraction_per_group = 1, 
  norm.method = "none", 
  p.adjust.method = "BH",
  p.threshold = 0.01,
  fc.threshold = log2(1.2), 
  keep.cols = c("Protein.Group", 
   "Genes", 
   "First.Protein.Description"), 
  rename.output = c(estimate = "log2.fc"))
}
}
