% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_grouped.R
\name{analyze_data_grouped}
\alias{analyze_data_grouped}
\title{Peptide-based limma analysis}
\usage{
analyze_data_grouped(
  data_processed,
  conditions = c("Control", "Treatment"),
  p.adjust.method = "BH",
  p.threshold = 0.01,
  fc.threshold = log2(1.2),
  eBayes.trend = TRUE,
  keep.cols,
  rename.output = c(estimate = "log2.fc"),
  full.output = F,
  silent = F
)
}
\arguments{
\item{data_processed}{output list or tibble from process_data_grouped()}

\item{conditions}{character vector specifying conditions to compare}

\item{p.adjust.method}{method for p.adjust() (see p.adjust.methods)}

\item{p.threshold}{p-value threshold after adjustment}

\item{fc.threshold}{log2. fold-change threshold}

\item{eBayes.trend}{abundance-dependent variance estimation in the 
limma::eBayes function (default = TRUE)}

\item{keep.cols}{Which columns should be copied from data_raw to 
data_peptides}

\item{rename.output}{named vector of columns to rename in the output (e.g. 
#' c(name = "new_name"))}

\item{full.output}{Include full limma data in output?}

\item{silent}{Turn off feedback from the function?}
}
\value{
tibble of limma results of full list of limma data incl. results
}
\description{
Peptide-based limma analysis
}
\examples{
\dontrun{
data_limma <- analyze_data_grouped(data_processed,
  conditions = c("Control", "Treatment"),
  p.adjust.method = "BH",
  p.threshold = 0.01,
  fc.threshold = log2(1.2), 
  keep.cols = c("Protein.Group", 
   "Genes", 
   "First.Protein.Description"), 
  rename.output = c(estimate = "log2.fc"))
}
}
