% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_annotation.R
\name{annotate_results_gw}
\alias{annotate_results_gw}
\title{Add protein annotations to group-wise PELSA results}
\usage{
annotate_results_gw(
  data,
  col_names = c(Peptide.Id = "Peptides", Protein.Description =
    "First.Protein.Description", log2.fc = "Log2FC", p.value = "P.Value"),
  fields = c("length", "ft_binding", "xref_pdb_full", "sequence"),
  max.query = 100,
  taxon_id = NULL,
  data_UniProt = NULL,
  add_anno2peptides = F,
  keep.peptide_length = T
)
}
\arguments{
\item{data}{results list or peptides data frame from analyze_data_grouped()}

\item{col_names}{named vector specifying which columns should be renamed}

\item{fields}{UniProt fields to be downloaded (see UniProt_fields())}

\item{max.query}{maximum number of proteins per UniProt query}

\item{taxon_id}{for single species data, taxon id to speed up UniProt 
download}

\item{data_UniProt}{predownloaded UniProt annotation data}

\item{add_anno2peptides}{Should protein annotations be added tot the 
peptides data frame?}

\item{keep.peptide_length}{Should the peptide length be kept in the data?}
}
\description{
Add protein annotations to group-wise PELSA results
}
