% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_annotation.R
\name{classify_peptide_protein_hits}
\alias{classify_peptide_protein_hits}
\title{classify protein and peptide hits in group-wise PELSA experiment}
\usage{
classify_peptide_protein_hits(
  data_peptides,
  p.value_column = "p.value",
  fc_column = "log2.fc",
  p.threshold.protein = 0.001,
  fc.threshold.protein = 0,
  p.threshold.peptide = 0.1,
  fc.threshold.peptide = 0.3,
  protein.id_column = "Protein.Group"
)
}
\arguments{
\item{data_peptides}{}

\item{p.value_column}{}

\item{fc_column}{}

\item{p.threshold.protein}{}

\item{fc.threshold.protein}{}

\item{p.threshold.peptide}{}

\item{fc.threshold.peptide}{}

\item{protein.id_column}{}
}
\description{
classify protein and peptide hits in group-wise PELSA experiment
}
