% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dose_response.R
\name{.fit_LL.4}
\alias{.fit_LL.4}
\title{Fitting function wrapper around drc::drm(..., fct = drc::LL.4()) used by 
map(data_fit, this_function) in analyze_data_dose_response()}
\usage{
.fit_LL.4(x)
}
\arguments{
\item{x}{data.frame object with columns log10.Concentration and Quant}
}
\value{
tibble containing fit parameters
}
\description{
Fitting function wrapper around drc::drm(..., fct = drc::LL.4()) used by 
map(data_fit, this_function) in analyze_data_dose_response()
}
