% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hello_I_want_to_analyse_my_PELSA_data}
\alias{.hello_I_want_to_analyse_my_PELSA_data}
\title{Helper function to get started}
\usage{
.hello_I_want_to_analyse_my_PELSA_data()
}
\value{
simply prints code to copy and use
}
\description{
Helper function to get started
}
\examples{
\dontrun{
.hello_I_want_to_analyse_my_PELSA_data()
}
}
