% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.longest_common_prefix}
\alias{.longest_common_prefix}
\title{Identify the longest common prefix of a character vector}
\usage{
.longest_common_prefix(strs)
}
\arguments{
\item{strs}{character vector}
}
\value{
string of longest common prefix ("" if nothing in common)
}
\description{
Identify the longest common prefix of a character vector
}
\examples{
.longest_common_prefix(c("prefix_A", "prefix_B"))
}
