% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.translate_concentrations}
\alias{.translate_concentrations}
\title{Translate numeric concentrations to strings with unit}
\usage{
.translate_concentrations(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
character vector
}
\description{
Translate numeric concentrations to strings with unit
}
\examples{
  .translate_concentrations(c(0, 1, 2e-3, 5e-6))
}
