% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_data_wide}
\alias{import_data_wide}
\title{Import wide-format data and (optionally) prepare the processing command}
\usage{
import_data_wide(
  file,
  data.origin = c("diann", "other"),
  experiment.type = c("grouped", "dose_response"),
  silent = F,
  ...
)
}
\arguments{
\item{file}{file location (select when empty)}

\item{data.origin}{by which software the data was searched}

\item{experiment.type}{whether the experimental design was group-wise ("grouped") 
or dose-response ("dose_response")}

\item{silent}{Turn off feedback from the function?}

\item{...}{further arguments if data.origin is "other" like <sample_names>}
}
\value{
A tibble of the imported data
}
\description{
Import wide-format data and (optionally) prepare the processing command
}
\examples{
import_data_wide("Data/report.pr_matrix.tsv", 
  data.origin = "diann", 
  experiment.type = "grouped")
}
