% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_annotation.R
\name{mutate_peptide_position}
\alias{mutate_peptide_position}
\title{Add peptide position based on stripped sequence column}
\usage{
mutate_peptide_position(
  x,
  sequences = c(A = "ABC"),
  sequence.col = "Stripped.Sequence",
  protein.col = "Protein.Group",
  keep.peptide_length = T,
  .after = sequence.col
)
}
\arguments{
\item{x}{tibble containing sequence and protein Id columns}

\item{sequences}{named vector containing protein sequences}

\item{sequence.col}{column containing peptide sequence}

\item{protein.col}{column containing protein Ids}

\item{keep.peptide_length}{Should the peptide length column be kept in the 
data?}

\item{.after}{position of the sequence position columns}
}
\description{
Add peptide position based on stripped sequence column
}
