% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dose_response.R
\name{prefilter_data_dose_response}
\alias{prefilter_data_dose_response}
\title{Prefilter dose-response data based on min. change at specified concentrations}
\usage{
prefilter_data_dose_response(
  data_processed,
  filter.thresholds = list(`1` = c(0.7, 1.3), `2` = c(1, 1))
)
}
\arguments{
\item{data_processed}{output list or tibble from process_data_dose_response()}

\item{filter.thresholds}{list of thresholds to filter peptides in the format 
list("rank of concentration" = c(ratio threshold stabilized, ratio threshold 
destabilized) with "1" = (1.) highest concentration etc.}
}
\value{
list or tibble of prefiltered peptide data for 
analyze_data_dose_response()
}
\description{
Prefilter dose-response data based on min. change at specified concentrations
}
\examples{
\dontrun{
data_processed <- prefilter_data_dose_response(
  data_processed = data_processed,
  filter.thresholds = list("1" = c(0.7, 1.3), 
    "2" = c(1, 1)))
}
}
