% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dose_response.R
\name{process_data_dose_response_m}
\alias{process_data_dose_response_m}
\title{Process the imported DIA-NN precursor data with name matching for inputs}
\usage{
process_data_dose_response_m(
  data_raw,
  sample_names,
  sample_concentrations,
  sample_replicates,
  peptide.column = "Stripped.Sequence",
  sample.column = "Run",
  quant.column = "Precursor.Normalised",
  filter_rows,
  min_fraction_per_rep = 1,
  min_reps_per_peptide = 3,
  scale_by_0 = T,
  silent = F
)
}
\arguments{
\item{data_raw}{tibble in long or wide format (e.g. from 
import_data_long()/_wide())}

\item{sample_names}{a list with the arguments pattern and (optional) replace}

\item{sample_concentrations}{a list with the arguments pattern and (optional) replace}

\item{sample_replicates}{a list with the arguments pattern and (optional) replace}

\item{peptide.column}{column containing peptide IDs to which the data should 
be summarized/collapse (by sum())}

\item{sample.column}{column containing sample names (when working 
with long format data)}

\item{quant.column}{column containing quantitative information (when working 
with long format data)}

\item{filter_rows}{tidy expression to filter the data e.g. filter_rows = 
Lib.Q.Value < 0.01}

\item{min_fraction_per_rep}{minimum fraction of values per replicate to use 
(e.g. 0.8 = min. 4 of 5 concentrations with valid value)}

\item{min_reps_per_peptide}{minimum number of complete replicates (see 
min_fraction_per_rep) to keep a peptide (default = 3; when using 4 
replicates)}

\item{scale_by_0}{Scale data by values at concentration 0?}

\item{silent}{Turn off feedback from the function?}
}
\value{
list of processed peptide data ($data_processed) and used raw data 
($data_raw)
}
\description{
Process the imported DIA-NN precursor data with name matching for inputs
}
\examples{
# see import_data_long()/_wide() output 
}
