% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_grouped.R
\name{process_data_grouped}
\alias{process_data_grouped}
\title{Process the imported DIA-NN precursor data}
\usage{
process_data_grouped(
  data_raw,
  sample_names,
  sample_groups,
  peptide.column = "Stripped.Sequence",
  sample.column = "Run",
  quant.column = "Precursor.Normalised",
  filter_rows,
  min_fraction_per_group = 1,
  norm.method = "none",
  silent = F
)
}
\arguments{
\item{data_raw}{tibble in long or wide format (e.g. from 
import_data_long()/_wide())}

\item{sample_names}{named vector containing sample names as value and 
original sample names as names (specifies which samples to use and renames 
them)}

\item{sample_groups}{named vector containing group information as value and 
original sample names as names}

\item{peptide.column}{column containing peptide IDs to which the data should 
be summarized/collapse (by sum())}

\item{sample.column}{column containing sample names (when working 
with long format data)}

\item{quant.column}{column containing quantitative information (when working 
with long format data)}

\item{filter_rows}{tidy expression to filter the data e.g. filter_rows = 
Lib.Q.Value < 0.01}

\item{min_fraction_per_group}{minimum fraction of values per replicate to use 
(e.g. 0.75 = min. 3 of 4 replicates with valid value)}

\item{norm.method}{which method to use for normalization}

\item{silent}{Turn off feedback from the function?}
}
\value{
list of processed peptide data ($data_processed) and used raw data 
($data_raw)
}
\description{
Process the imported DIA-NN precursor data
}
\examples{
# see import_data_long()/_wide() output 
}
